+++ libkbfs/crypto_common.go
@@ -20,20 +20,6 @@ import (
 	"golang.org/x/crypto/nacl/secretbox"
 )
 
-// Belt-and-suspenders wrapper around crypto.rand.Read().
-func cryptoRandRead(buf []byte) error {
-	n, err := rand.Read(buf)
-	if err != nil {
-		return err
-	}
-	// This is truly unexpected, as rand.Read() is supposed to
-	// return an error on a short read already!
-	if n != len(buf) {
-		return UnexpectedShortCryptoRandRead{}
-	}
-	return nil
-}
-
 // CryptoCommon contains many of the function implementations need for
 // the Crypto interface, which can be reused by other implementations.
 type CryptoCommon struct {
@@ -50,7 +36,7 @@ func MakeCryptoCommon(codec kbfscodec.Codec) CryptoCommon {
 // MakeRandomTlfID implements the Crypto interface for CryptoCommon.
 func (c CryptoCommon) MakeRandomTlfID(isPublic bool) (tlf.ID, error) {
 	var idBytes [tlf.IDByteLen]byte
-	err := cryptoRandRead(idBytes[:])
+	err := kbfscrypto.RandRead(idBytes[:])
 	if err != nil {
 		return tlf.ID{}, err
 	}
@@ -70,7 +56,7 @@ func (c CryptoCommon) MakeRandomTlfID(isPublic bool) (tlf.ID, error) {
 // MakeRandomBranchID implements the Crypto interface for CryptoCommon.
 func (c CryptoCommon) MakeRandomBranchID() (BranchID, error) {
 	var id BranchID
-	err := cryptoRandRead(id.id[:])
+	err := kbfscrypto.RandRead(id.id[:])
 	if err != nil {
 		return BranchID{}, err
 	}
@@ -141,7 +127,7 @@ func (c CryptoCommon) MakeTLFReaderKeyBundleID(rkb *TLFReaderKeyBundleV3) (
 // MakeTemporaryBlockID implements the Crypto interface for CryptoCommon.
 func (c CryptoCommon) MakeTemporaryBlockID() (BlockID, error) {
 	var dh kbfshash.RawDefaultHash
-	err := cryptoRandRead(dh[:])
+	err := kbfscrypto.RandRead(dh[:])
 	if err != nil {
 		return BlockID{}, err
 	}
@@ -168,7 +154,7 @@ func (c CryptoCommon) VerifyBlockID(encodedEncryptedData []byte, id BlockID) err
 
 // MakeBlockRefNonce implements the Crypto interface for CryptoCommon.
 func (c CryptoCommon) MakeBlockRefNonce() (nonce BlockRefNonce, err error) {
-	err = cryptoRandRead(nonce[:])
+	err = kbfscrypto.RandRead(nonce[:])
 	return
 }
 
@@ -209,7 +195,7 @@ func (c CryptoCommon) MakeRandomTLFKeys() (
 		*keyPair.Private)
 
 	var data [32]byte
-	err = cryptoRandRead(data[:])
+	err = kbfscrypto.RandRead(data[:])
 	if err != nil {
 		return
 	}
@@ -223,7 +209,7 @@ func (c CryptoCommon) MakeRandomTLFKeys() (
 func (c CryptoCommon) MakeRandomTLFCryptKeyServerHalf() (
 	serverHalf kbfscrypto.TLFCryptKeyServerHalf, err error) {
 	var data [32]byte
-	err = cryptoRandRead(data[:])
+	err = kbfscrypto.RandRead(data[:])
 	if err != nil {
 		return kbfscrypto.TLFCryptKeyServerHalf{}, err
 	}
@@ -236,7 +222,7 @@ func (c CryptoCommon) MakeRandomTLFCryptKeyServerHalf() (
 func (c CryptoCommon) MakeRandomBlockCryptKeyServerHalf() (
 	serverHalf kbfscrypto.BlockCryptKeyServerHalf, err error) {
 	var data [32]byte
-	err = cryptoRandRead(data[:])
+	err = kbfscrypto.RandRead(data[:])
 	if err != nil {
 		return kbfscrypto.BlockCryptKeyServerHalf{}, err
 	}
@@ -282,7 +268,7 @@ func (c CryptoCommon) EncryptTLFCryptKeyClientHalf(
 	clientHalf kbfscrypto.TLFCryptKeyClientHalf) (
 	encryptedClientHalf EncryptedTLFCryptKeyClientHalf, err error) {
 	var nonce [24]byte
-	err = cryptoRandRead(nonce[:])
+	err = kbfscrypto.RandRead(nonce[:])
 	if err != nil {
 		return
 	}
@@ -312,7 +298,7 @@ func (c CryptoCommon) EncryptTLFCryptKeyClientHalf(
 
 func (c CryptoCommon) encryptData(data []byte, key [32]byte) (encryptedData, error) {
 	var nonce [24]byte
-	err := cryptoRandRead(nonce[:])
+	err := kbfscrypto.RandRead(nonce[:])
 	if err != nil {
 		return encryptedData{}, err
 	}
@@ -429,7 +415,7 @@ func (c CryptoCommon) padBlock(block []byte) ([]byte, error) {
 		return nil, err
 	}
 	if n != padLen {
-		return nil, UnexpectedShortCryptoRandRead{}
+		return nil, kbfscrypto.UnexpectedShortCryptoRandRead{}
 	}
 
 	return buf.Bytes(), nil
